+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add option to manually start the scan
 - Add configuration for tooltip
+- Don't refresh when regex is being changed
 
 ## v0.0.169 - 2020-02-25